backend/requirements.txt
aniso8601
attrs
azure-core
azure-cosmos
blinker
certifi
charset-normalizer
click
Flask
flask-restx
idna
importlib_resources
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
numpy
pandas
python-dateutil
python-dotenv
pytz
referencing
requests
rpds-py
six
typing_extensions
tzdata
urllib3
Werkzeug
flask-cors
chromadb

# === Agent + LLM Tools ===
aisuite==0.1.11
anthropic
docstring-parser
markdown
mistralai
openai
qrcode
tavily-python>=0.7.12
textstat
vertexai
langchain-core
langchain-openai
langgraph

# === Web Framework + API ===
fastapi
pydantic
pydantic[email]
python-dotenv
python-multipart
requests
sqlalchemy
uvicorn

# === Notebook Experience ===
ipywidgets
jupyter_server
azure-cosmos
nbclassic
notebook

# === Data Analysis / Display (Optional Enhancements) ===
duckdb
matplotlib
pandas
seaborn
tabulate
tinydb

# === Machine Learning / NLP (Optional Enhancements) ===
jinja2
psycopg2-binary
scikit-learn
Wikipedia

azure-cosmos
langchain-openai
langchain-azure-ai
langchain_community
pypdf
chromadb
sentence-transformers
google-genai

Flask
requests
pandas